replace the impl. of byte.ByteData with haxe.io.Bytes
This patch replace the implementation of byte.ByteData with Bytes of the standard API, intended to follow up the upstream bugfix.

NOTE: the "skip the bom" code is totally incorrect<sup>[1]</sup> and BOM won't be the matter, I just replace ByteData with Bytes. If BOM is actually the matter, I'll devise how to treat it.

[1]: Checking the code of the first letter is 239(0xEF) is nonsense because encoding of strings is platform-dependent, so it may be U+00EF "ï". The following code assumes the encoding of strings is UTF-32. Even in UTF-8 environment, every characters in U+FXXX has 0xEF in their first byte in UTF-8.